[feat] 여행목록 UI 초안3
